+++ src/test/java/roomescape/AdminPageControllerTest.java
@@ -1,0 +1,41 @@
+package roomescape;
+
+import io.restassured.RestAssured;
+import org.junit.jupiter.api.BeforeEach;
+import org.junit.jupiter.api.DisplayName;
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.boot.test.context.SpringBootTest.WebEnvironment;
+import org.springframework.boot.test.web.server.LocalServerPort;
+import org.springframework.test.annotation.DirtiesContext;
+
+@SpringBootTest(webEnvironment = WebEnvironment.RANDOM_PORT)
+@DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
+class AdminPageControllerTest {
+
+    @LocalServerPort
+    int serverPort;
+
+    @BeforeEach
+    public void beforeEach() {
+        RestAssured.port = serverPort;
+    }
+
+    @Test
+    @DisplayName("방탈출 어드민 메인 페이지 조회를 확인한다")
+    void showAdminMainPage() {
+        RestAssured.given().log().all()
+                .when().get("/admin")
+                .then().log().all()
+                .statusCode(200);
+    }
+
+    @Test
+    @DisplayName("방탈출 예약 관리 페이지 조회를 확인한다")
+    void showReservationPage() {
+        RestAssured.given().log().all()
+                .when().get("/admin/reservation")
+                .then().log().all()
+                .statusCode(200);
+    }
+}
